pyproject.toml
[tool.poetry]
name = "postgres-da-ai-agent"
version = "0.1.0"
description = ""
authors = ["benjaminbascary97 <benjamin.bascary@patagoniansys.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.1"
psycopg2-binary = "^2.9.9"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
pyautogen = "^0.1.12"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "postgres_da_ai_agent.main:main"